set colorspace
Meh, D3D11's support for this is so poor.

ID3D11VideoContext1::VideoProcessorSetStreamColorSpace1 provides slightly
better control, but is not available in mingw yet. Since it derives
from ID3D11VideoContext1::VideoProcessorSetStreamColorSpace1 provides slightly
better control, but is not available in mingw yet. Since it derives
from ID3D11VideoContext, which is huge, it's not too nice to declare
it manually either.